requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
flake8==7.3.0
future==1.0.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
load-dotenv==0.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
serializers==0.2.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
